PP-13733 Extract transform functions
